+++ apps/backend/backend_commit.c
@@ -285,7 +285,8 @@ candidate_commit(clicon_handle h,
 
      /* Optionally write (potentially modified) tree back to candidate */
      if (clicon_option_bool(h, "CLICON_TRANSACTION_MOD")){
-	 if ((ret = xmldb_put(h, candidate, OP_REPLACE, td->td_target, cbret)) < 0)
+	 if ((ret = xmldb_put(h, candidate, OP_REPLACE, td->td_target,
+			      clicon_username_get(h), cbret)) < 0)
 	     goto done;
 	 if (ret == 0)
 	     goto fail;
@@ -440,7 +441,8 @@ from_client_validate(clicon_handle h,
     }
     /* Optionally write (potentially modified) tree back to candidate */
     if (clicon_option_bool(h, "CLICON_TRANSACTION_MOD")){
-	if ((ret = xmldb_put(h, "candidate", OP_REPLACE, td->td_target, cbret)) < 0)
+	if ((ret = xmldb_put(h, "candidate", OP_REPLACE, td->td_target,
+			     clicon_username_get(h), cbret)) < 0)
 	    goto done;
 	goto ok;
     }